refactor: survey 도메인의 위치를 root모듈로 변경한다
