+++ src/main/kotlin/domain/survey/ChoiceFormQuestionType.kt
@@ -1,0 +1,6 @@
+package domain.survey
+
+enum class ChoiceFormQuestionType {
+    TENDENCY,
+    CUSTOM,
+}
