+++ src/main/kotlin/domain/survey/QuestionType.kt
@@ -1,0 +1,6 @@
+package domain.survey
+
+enum class QuestionType {
+    CHOICE,
+    SHORT,
+}
