+++ src/main/kotlin/domain/survey/ShortFormQuestionType.kt
@@ -1,0 +1,7 @@
+package domain.survey
+
+enum class ShortFormQuestionType {
+    STRENGTH,
+    WEAKNESS,
+    CUSTOM,
+}
