+++ src/main/kotlin/domain/survey/SurveyBookmark.kt
@@ -1,0 +1,12 @@
+package domain.survey
+
+import javax.persistence.Column
+import javax.persistence.Embeddable
+import javax.persistence.JoinColumn
+
+@Embeddable
+class SurveyBookmark(
+    @Column(name = "bookmarked_survey_id")
+    @JoinColumn(name = "survey_id", nullable = false)
+    val surveyId: Long
+)
