+++ src/main/kotlin/domain/survey/SurveyRepository.kt
@@ -1,0 +1,5 @@
+package domain.survey
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface SurveyRepository : JpaRepository<Survey, Long>
