+++ src/main/kotlin/domain/survey/exception/DuplicatedOrderException.kt
@@ -1,0 +1,6 @@
+package domain.survey.exception
+
+class DuplicatedOrderException internal constructor(duplicated: Int, orders: HashSet<Int>) :
+    RuntimeException(
+        "Duplicated order detected duplicated \"$duplicated\" ordinary \"$orders\""
+    )
