--- survey/build.gradle
@@ -1,12 +1,0 @@
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation project(":core:data")
-    implementation project(":core:time")
-
-    implementation "org.springframework.boot:spring-boot-starter-data-jpa"
-
-    testRuntimeOnly "com.h2database:h2"
-}
