--- survey/src/main/kotlin/me/nalab/survey/domain/feedback/ChoiceFormQuestionFeedback.kt
@@ -1,17 +1,0 @@
-package me.nalab.survey.domain.feedback
-
-import javax.persistence.*
-
-@Entity
-class ChoiceFormQuestionFeedback(
-    id: Long,
-    formQuestionId: Long,
-    isRead: Boolean = false,
-    bookmark: Bookmark = Bookmark.impossible(),
-    feedback: Feedback,
-
-    @ElementCollection
-    @Column(name = "selects")
-    @CollectionTable(name = "selects", joinColumns = [JoinColumn(name = "form_feedback_id")])
-    private val selectedChoiceIds: MutableSet<Long>,
-) : FormQuestionFeedbackable(id, formQuestionId, isRead, bookmark, feedback)
