--- survey/src/main/kotlin/me/nalab/survey/domain/survey/ChoiceFormQuestionType.kt
@@ -1,6 +1,0 @@
-package me.nalab.survey.domain.survey
-
-enum class ChoiceFormQuestionType {
-    TENDENCY,
-    CUSTOM,
-}
