--- survey/src/main/kotlin/me/nalab/survey/domain/survey/QuestionType.kt
@@ -1,6 +1,0 @@
-package me.nalab.survey.domain.survey
-
-enum class QuestionType {
-    CHOICE,
-    SHORT,
-}
