--- survey/src/main/kotlin/me/nalab/survey/domain/survey/ShortFormQuestion.kt
@@ -1,16 +1,0 @@
-package me.nalab.survey.domain.survey
-
-import javax.persistence.*
-
-@Entity
-class ShortFormQuestion(
-    id: Long,
-    title: String,
-    order: Int,
-    questionType: QuestionType,
-    survey: Survey,
-
-    @Enumerated(EnumType.STRING)
-    @Column(name = "short_form_question_type")
-    val shortFormQuestionType: ShortFormQuestionType,
-) : FormQuestionable(id, title, order, questionType, survey)
