--- survey/src/main/kotlin/me/nalab/survey/domain/survey/ShortFormQuestionType.kt
@@ -1,7 +1,0 @@
-package me.nalab.survey.domain.survey
-
-enum class ShortFormQuestionType {
-    STRENGTH,
-    WEAKNESS,
-    CUSTOM,
-}
