--- survey/src/main/kotlin/me/nalab/survey/domain/survey/SurveyRepository.kt
@@ -1,5 +1,0 @@
-package me.nalab.survey.domain.survey
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface SurveyRepository : JpaRepository<Survey, Long>
