--- survey/src/main/kotlin/me/nalab/survey/domain/survey/exception/DuplicatedOrderException.kt
@@ -1,6 +1,0 @@
-package me.nalab.survey.domain.survey.exception
-
-class DuplicatedOrderException internal constructor(duplicated: Int, orders: HashSet<Int>) :
-    RuntimeException(
-        "Duplicated order detected duplicated \"$duplicated\" ordinary \"$orders\""
-    )
