--- survey/src/main/kotlin/me/nalab/survey/domain/survey/value/ImageUrl.kt
@@ -1,15 +1,0 @@
-package me.nalab.survey.domain.survey.value
-
-import javax.persistence.Column
-import javax.persistence.Embeddable
-
-@Embeddable
-class ImageUrl(
-    @Column(name = "image_url", nullable = false)
-    val value: String,
-) {
-
-    companion object {
-        fun empty(): ImageUrl = ImageUrl("")
-    }
-}
