--- survey/src/main/kotlin/me/nalab/survey/domain/survey/value/Job.kt
@@ -1,15 +1,0 @@
-package me.nalab.survey.domain.survey.value
-
-import javax.persistence.Column
-import javax.persistence.Embeddable
-
-@Embeddable
-class Job(
-    @Column(name = "job", nullable = false)
-    val value: String,
-) {
-
-    companion object {
-        fun empty(): Job = Job("")
-    }
-}
