--- survey/src/main/kotlin/me/nalab/survey/domain/survey/value/Position.kt
@@ -1,15 +1,0 @@
-package me.nalab.survey.domain.survey.value
-
-import javax.persistence.Column
-import javax.persistence.Embeddable
-
-@Embeddable
-class Position(
-    @Column(name = "position", nullable = false)
-    val value: String,
-) {
-
-    companion object {
-        fun empty(): Position = Position("")
-    }
-}
